Return type of sizeof() might be int or long.
git-svn-id: https://svn.aros.org/svn/aros/trunk@772 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ config/machine.c
@@ -111,7 +111,7 @@ int main (void)
 
     printf ("#define AROS_STACK_GROWS_DOWNWARDS %d /* Stack direction */\n", (adr2 < adr1));
     printf ("#define AROS_BIG_ENDIAN            %d /* Big or little endian */\n", (*first_byte == 0x11));
-    printf ("#define AROS_SIZEOFULONG           %d /* Size of an ULONG */\n", sizeof (ULONG));
+    printf ("#define AROS_SIZEOFULONG           %ld /* Size of an ULONG */\n", (unsigned long)sizeof (ULONG));
     printf ("#define AROS_WORDALIGN             %d /* Alignment for WORD */\n",   wordalign);
     printf ("#define AROS_LONGALIGN             %d /* Alignment for LONG */\n",   longalign);
     printf ("#define AROS_PTRALIGN              %d /* Alignment for PTR */\n",    ptralign);
